+++ app/Models/BillingInvoice.php
@@ -28,4 +28,9 @@ final class BillingInvoice extends Model
     {
         return $this->belongsTo(Subscription::class);
     }
+
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
 }