+++ app/Models/SubscriptionEvent.php
@@ -26,4 +26,9 @@ final class SubscriptionEvent extends Model
     {
         return $this->belongsTo(Subscription::class);
     }
+
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
 }